Move drag module from widgets to input
