+++ src/widgets/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod button;
 pub mod scroll;
-pub mod drag;
 pub mod list;
 pub mod slider;
 pub mod edit_text;
