+++ src/widgets/slider.rs
@@ -5,11 +5,11 @@ use cassowary::strength::*;
 use layout::constraint::*;
 use layout::LayoutUpdated;
 use input::mouse::ClickEvent;
+use input::drag::{DragEvent, DragState};
 use event::{EventHandler, EventArgs};
 use widget::{WidgetBuilder, WidgetRef};
 use widget::property::Property;
 use widget::property::states::*;
-use widgets::drag::{DragEvent, DragState};
 use draw::rect::{RectState, RectStyle};
 use draw::ellipse::{EllipseState, EllipseStyle};
 use geometry::{RectExt, Point};
